+++ discord/raw_models.py
@@ -70,6 +70,7 @@ if TYPE_CHECKING:
         ThreadMembersUpdateEvent,
         ThreadUpdateEvent,
         TypingEvent,
+        VoiceChannelStartTimeUpdateEvent,
         VoiceChannelStatusUpdateEvent,
         VoiceServerUpdateEvent,
         VoiceStateEvent,
@@ -98,7 +99,9 @@ __all__ = (
     "RawSoundboardSoundDeleteEvent",
     "RawVoiceServerUpdateEvent",
     "RawVoiceStateUpdateEvent",
+    "RawVoiceChannelStartTimeUpdateEvent",
     "RawMemberUpdateEvent",
+    "ChannelInfo",
 )
 
 
@@ -1030,3 +1033,54 @@ class RawMemberUpdateEvent(_RawReprMixin):
         self.data: MemberUpdateEvent = data
         self.cached_member: Member | None = None
         self.member: Member = member
+
+
+class RawVoiceChannelStartTimeUpdateEvent(_RawReprMixin):
+    """Represents the payload for an :func:`on_raw_voice_channel_start_time_update` event.
+
+    .. versionadded:: 2.9
+
+    Attributes
+    ----------
+    id: :class:`int`
+        The channel ID where the voice channel start time update originated from.
+    guild_id: :class:`int`
+        The guild ID where the voice channel start time update originated from.
+    voice_start_time: Optional[:class:`datetime.datetime`]
+        The new new voice channel start time.
+    data: :class:`dict`
+        The raw data sent by the `gateway <https://docs.discord.com/developers/events/gateway-events-events#voice-channel-start-time-update>`__.
+    """
+
+    __slots__ = ("id", "guild_id", "voice_start_time", "data")
+
+    def __init__(self, data: VoiceChannelStartTimeUpdateEvent) -> None:
+        self.id: int = int(data["id"])
+        self.guild_id: int = int(data["guild_id"])
+        self.voice_start_time: datetime.datetime | None = (
+            datetime.datetime.fromtimestamp(data["voice_start_time"], tz=datetime.UTC)
+            if data.get("voice_start_time")
+            else None
+        )
+        self.data: VoiceChannelStartTimeUpdateEvent = data
+
+
+class ChannelInfo(_RawReprMixin):
+    """Represents the gateway response to a request for channel information.
+
+    .. versionadded:: 2.9
+
+    Attributes
+    ----------
+    id: :class:`int`
+        The ID of the channel this info is associated with.
+    status: :class:`str` | None
+        The voice channel status.
+    voice_start_time: :class:`int` | None
+        The Unix timestamp (in seconds) of when the voice session started.
+    """
+
+    def __init__(self, data):
+        self.id: int = int(data["id"])
+        self.status: str | None = data.get("status", utils.MISSING)
+        self.voice_start_time: int | None = data.get("voice_start_time", utils.MISSING)